+++ src/pages/CSS/reservationCSS.module.css
@@ -1,0 +1,35 @@
+body {
+    font-family: "Montserrat", sans-serif;
+    color: black;
+    font-size: 1em;
+    text-align: center;
+}
+
+#wrapper {
+    margin: 0 auto;
+    width: 960px;
+    text-align: left;
+    color: #000;
+    background-color: rgb(237, 250, 255);
+    margin-top: 50px;
+    padding: 20px;
+    padding-top: 40px;
+}
+
+h1 {
+    text-align: center;
+    background-color: bisque;
+    padding: 20px;
+    margin-left: 30px;
+    margin-right: 30px;
+}
+
+h2 {
+    text-align: center;
+    padding: 30px;
+}
+
+#row {
+    text-align: center;
+    margin: 30px;
+}
